+++ VERSION_HISTORY.markdown
@@ -7,6 +7,7 @@
 * Large logging cleanup and bug fix - upshot is `--no-color` works again
 * Adds a full `punic update` to test suite.
 * Fixes punic on Python 3.5 which was a little bit broken.
+* Fixes #8 - show a useful error if user doesn't have graphviz and/or pydotplus installed.
 
 ## 0.1.17